Prediction File
This File includes the Prediction part of the repository in which the prediction has been made about the survival on the basis of labels which is mentioned in the csv file. The model has been trained and tested on two machine Learning model such as: Linear Regression and Decision Tree Classifier.